Can now remove biotopes and also create a 2d without spawning a landscape
There is a bug where if u delete a biotope and then close the plugin and reopen it will crash.
This is due to hard coding that some slots of the settings array should remove old settings, but since these have been removed, it will be out of bounds.

Correct solution would be to add a button to add "preset settings / assets"